refactor: support headless auth in windows
